requirements.txt
absl-py==2.0.0
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
Babel==2.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.4
boto3==1.28.16
cachetools==5.3.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.0
Click==7.0
colorama==0.4.3
comm==0.1.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
debtcollector==2.0.0
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.6
django-cors-headers==4.2.0
django-mysql==4.11.0
django-rest-framework==0.1.0
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.16
entrypoints==0.3
executing==1.2.0
filelock==3.0.12
flatbuffers==23.5.26
fonttools==4.42.1
frozendict==2.3.8
fuzzywuzzy==0.18.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
html5lib==1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.1.0
incremental==16.10.1
ipykernel==6.25.0
ipython==8.12.2
iso8601==0.1.12
jedi==0.19.0
Jinja2==2.10.1
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Levenshtein==0.23.0
libclang==16.0.6
lxml==4.9.3
macaroonbakery==1.3.1
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
monotonic==1.5
more-itertools==4.2.0
msgpack==0.6.2
multitasking==0.0.11
nest-asyncio==1.5.7
netaddr==0.7.19
netifaces==0.10.4
numpy==1.24.3
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
oslo.config==8.0.2
oslo.context==3.0.2
oslo.i18n==4.0.1
oslo.log==4.1.1
oslo.serialization==3.1.1
oslo.utils==4.1.1
packaging==23.1
pandas==2.0.3
pandas-datareader==0.10.0
parso==0.8.3
pbr==5.4.5
peewee==3.16.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.5.1
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.15.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-Levenshtein==0.23.0
python-novnc==1.0.0
pytz==2023.3.post1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.0
rapidfuzz==3.4.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3986==1.3.2
roman==2.0.0
rsa==4.9
s3transfer==0.6.1
scikit-learn==1.3.1
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
shortuuid==1.0.11
simplejson==3.16.0
six==1.14.0
sos==4.3
soupsieve==2.5
sqlparse==0.4.4
ssh-import-id==5.10
stack-data==0.6.2
stevedore==1.32.0
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tornado==6.3.2
traitlets==5.9.0
Twisted==18.9.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==27.11.2
ufw==0.36
unattended-upgrades==0.1
urllib3==2.0.5
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.6
webencodings==0.5.1
websockify==0.9.0
Werkzeug==3.0.0
wrapt==1.14.1
yfinance==0.2.30
zipp==3.17.0
zope.interface==4.7.1
